Add file utilities

+++ src/core/Reflection.h
@@ -28,7 +28,7 @@ namespace  hikari {
         BxDF(MaterialType type) : type(type) {}
         //bool MatchesFlags(MaterialType t) const { return (type & t) == type; }
         virtual Vector3 f(const Vector3 &wo, const Vector3 &wi) const = 0;
-        virtual Vector3 Sample_f(const Vector3 &wo, Vector3 *wi, const Vector2 &sample, double *pdf, BxDFType sampledType) const;
+        virtual Vector3 Sample_f(const Vector3 &wo, Vector3 *wi, const Vector2 &sample, float *pdf, BxDFType sampledType) const;
         virtual Vector3 rho(const Vector3 &wo, int nSamples, const Vector2 *samples) const;
         virtual Vector3 rho(int nSamples, const Vector2 *samples1, const Vector2 *samples2) const;
         virtual double Pdf(const Vector3 &wo, const Vector3 &wi) const;
